requirements.txt
certifi==2020.12.5
h11==0.12.0
httpcore==0.13.2
httpx==0.18.1
idna==3.1
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
